# src/main/resources/OpenAPI.yaml
openapi: 3.0.1
info:
  title: MS Product
  description: Microsserviço de produtos da barbearia
  version: 1.0.0
servers:
  - url: http://localhost:8084/api/barbershop
paths:
  /product/{id}:
    get:
      tags:
        - Product
      summary: Buscar produto por ID
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
    put:
      tags:
        - Product
      summary: Atualizar produto
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    delete:
      tags:
        - Product
      summary: Excluir produto
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
  /product:
    get:
      tags:
        - Product
      summary: Listar produtos
      operationId: findAll
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableDTO'
    post:
      tags:
        - Product
      summary: Cadastrar produto
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
components:
  schemas:
    ProductDTO:
      required:
        - name
        - quantity
        - value
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        value:
          type: number
        quantity:
          type: integer
          format: int64
    ProductResponse:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        value:
          type: number
        quantity:
          type: integer
          format: int64
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageableDTO:
      type: object
      properties:
        numberOfElements:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        paymentsList:
          type: array
          items:
            $ref: '#/components/schemas/ProductDTO'